requirements.txt
# Core network monitoring dependencies
psutil>=5.9.0
speedtest-cli>=2.1.3
pywifi>=1.1.12
pystray>=0.19.4
Pillow>=10.0.0
cairosvg>=2.7.1
PyInstaller>=6.3.0
pywin32>=306
inno-setup>=6.2.2

# Web dashboard dependencies
flask>=2.0.1
flask-sqlalchemy>=2.5.1
flask-cors>=3.0.10
SQLAlchemy>=1.4.23
requests>=2.26.0